requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
scikit-learn==1.5.2
numpy==2.0.2
pandas==2.2.3
joblib==1.4.2
python-multipart==0.0.9
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.2
clearml==1.16.3
boto3==1.35.36
dvc[s3]==3.59.0
requests==2.32.3
streamlit==1.39.0
